requirements.txt
python-gvm
fastapi